# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/password": {
      "patch": {
        "description": "Altera a senha do próprio usuário. Apenas o próprio usuário pode alterar sua senha.",
        "operationId": "ChangeUserPasswordController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário cuja senha será alterada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Senha antiga e nova senha",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "example": "oldPassword123",
                    "description": "Senha atual do usuário",
                    "minLength": 6
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "newPassword123",
                    "description": "Nova senha do usuário",
                    "minLength": 6
                  }
                },
                "required": ["oldPassword", "newPassword"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password changed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Dados inválidos" },
          "401": {
            "description": "Senha antiga incorreta ou token JWT inválido"
          },
          "403": {
            "description": "Usuário não tem permissão para alterar esta senha"
          },
          "404": { "description": "Usuário não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Alterar senha do usuário",
        "tags": ["User Profile"]
      }
    },
...
    "/categories/{id}": {
      "patch": {
        "description": "Edita uma categoria existente (requer role MANAGER+)",
        "operationId": "EditCategoryController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da categoria a ser editada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditCategoryRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Categoria editada com sucesso" },
          "400": { "description": "Dados inválidos ou categoria na lixeira" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Categoria não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Editar categoria",
        "tags": ["Categories"]
      },
      "delete": {
        "description": "Deleta permanentemente uma categoria do sistema (requer role MANAGER+)",
        "operationId": "DeleteCategoryController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da categoria a ser deletada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Categoria deletada com sucesso" },
          "400": { "description": "Categoria não encontrada" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Categoria não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Deletar categoria",
        "tags": ["Categories"]
      }
    },
...
    "/locations": {
      "post": {
        "description": "Cria uma nova localização no sistema (requer role MANAGER+)",
        "operationId": "CreateLocationController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateLocationRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Localização criada com sucesso" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Criar localização",
        "tags": ["Locations"]
      },
      "get": {
        "description": "Retorna uma lista paginada de localizações disponíveis no sistema",
        "operationId": "FetchLocationsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "example": 1, "type": "number" }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir localizações deletadas (na lixeira)",
            "schema": { "example": false, "type": "boolean" }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": { "example": true, "type": "boolean" }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Termo de busca para filtrar localizações por nome, código ou descrição",
            "schema": { "example": "Bloco A", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de localizações retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "locations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocationResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Parâmetros inválidos" }
        },
        "summary": "Buscar localizações",
        "tags": ["Locations"]
      }
    },
    "/locations/{id}": {
      "patch": {
        "description": "Edita uma localização existente (requer role MANAGER+)",
        "operationId": "EditLocationController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da localização a ser editada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditLocationRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Localização editada com sucesso" },
          "400": { "description": "Dados inválidos ou localização na lixeira" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Localização não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Editar localização",
        "tags": ["Locations"]
      },
      "delete": {
        "description": "Deleta permanentemente uma localização do sistema (requer role MANAGER+)",
        "operationId": "DeleteLocationController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da localização a ser deletada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Localização deletada com sucesso" },
          "400": { "description": "Localização não encontrada" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Localização não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Deletar localização",
        "tags": ["Locations"]
      }
    },
...
  },
  "info": {
    "title": "IFAL Arapiraca Backend API",
    "description": "API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "EditCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria",
            "example": "Climatização"
          },
          "description": {
            "type": "string",
            "description": "Descrição da categoria",
            "example": "Problemas relacionados a ar condicionado, ventilação e climatização"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status ativo/inativo da categoria",
            "example": true
          }
        }
      },
      "CreateLocationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da localização",
            "example": "Bloco A - Sala 201"
          },
          "description": {
            "type": "string",
            "description": "Descrição da localização",
            "example": "Sala de aula localizada no primeiro andar do Bloco A"
          },
          "code": {
            "type": "string",
            "description": "Código identificador da localização",
            "example": "BLA-201"
          }
        },
        "required": ["name", "description", "code"]
      },
      "EditLocationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da localização",
            "example": "Bloco A - Sala 201"
          },
          "code": {
            "type": "string",
            "description": "Código identificador da localização",
            "example": "BLA-201"
          },
          "description": {
            "type": "string",
            "description": "Descrição da localização",
            "example": "Sala de aula localizada no primeiro andar do Bloco A"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status ativo/inativo da localização",
            "example": true
          }
        }
      },
...
    }
  }
}